Added authentication for delete posts

+++ part4/blog-list/controllers/blogs.js
@@ -11,10 +11,12 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const token = request.token
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
+  const decodedToken = jwt.verify(token, process.env.SECRET, (err, decoded) => {
+    if (err) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    return decoded
+  })
   const user = await User.findById(decodedToken.id)
 
   const blog = await new Blog(request.body)
@@ -26,10 +28,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog
-    .findByIdAndRemove(request.params.id)
+  const token = request.token
+  const decodedToken = jwt.verify(token, process.env.SECRET, (err, decoded) => {
+    if (err) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    return decoded
+  })
+  const user = await User.findById(decodedToken.id)
+  const author = await Blog.findById(request.params.id)
 
-  response.status(204).end()
+  if (user.username === author.author && author !== undefined) {
+    await Blog
+      .findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+
+  }
+  return response.status(400).json({ error: 'only the author can delete the blog' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
